Round system and power-up cards kind of there
There are 3 power-up cards where 2 should be working but they don't. After the first round you are able to choose a power-up card and you move to the second round but the power-up card effect does not work and when you have played the cards from the deck to the base the game crashes (might be because of auto solve idk).